config2MQTT-watchdog/requirements.txt
paho-mqtt
watchdog